restructure inventory components
