+++ packages/client/src/components/inventory-manager.component.tsx
@@ -1,6 +1,6 @@
 import { Inventory, PartyInventory } from "@pf2e-inventory/shared"
 import AddItemDrawer from "./drawer/add-item-drawer.component"
-import InventoryGrid from "./grid/inventory-grid.component"
+import InventoryGrid from "./inventory-grid.component"
 import { useEffect, useState } from "react"
 import useWebSocket, { ReadyState } from "react-use-websocket"
 import { InventoryContext } from "../lib/inventory.context"
